í´§ Fixed ESLint missconfiguration and upgraded to Next.js 13! íº
We had a pesky ESLint issue that was causing our build to fail, but thanks to some careful debugging and a bit of elbow grease, we were able to resolve the issue and get our project back on track.

In addition, we have upgraded to Next.js 13, which brings a number of exciting new features and performance improvements. Our app is now running faster and smoother than ever before!

Huge thanks to our intrepid team of developers who worked tirelessly to get this fix out the door. You guys rock! í´