🎲 Added Entity interface for Monopoly game 🎲
We have introduced a new Entity interface that describes either a player or the bank in a Monopoly game. This interface includes properties for the entity's name, their current bank balance, and an array of the property IDs that they own.

This interface will be instrumental in keeping track of game state and ensuring that all players have a fair and enjoyable experience.

Thanks to the clever minds on our development team for coming up with this solution! Let the games begin! 🎉