🌐 Implemented REST API for property and entity management in Monopoly game 🌐
We have implemented a REST API for our Monopoly game that allows us to manage the property and entity JSON files. This API includes GET routes for retrieving property data, property data by ID, and entity data, as well as a POST route for creating new entities (players).

This API will be crucial for building the frontend of our Monopoly game and allowing players to interact with the game data.

Huge thanks to the hardworking development team for bringing this feature to fruition. Let the property buying and entity creation begin! 🎲